part4.21*: bloglist expansion, step9

+++ controllers/blogs.js
@@ -78,8 +78,18 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+  if(user._id.toString() === blogToDelete.user.toString()){
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+  } else{
+    return response.status(401).json({ error: 'wrong user attmpt' })
+  }    
 })
 
 blogsRouter.put('/:id', async (request, response) => {
